WIP: Single-threaded MattesMutualInformation
Hans Johnson noticed that that the Multi-threaded
MattesMutualInformation metric was actually slower with
multiple threads than with a single thread.

This adds a single-threaded version of the filter.

Due to the structure of the Registration Framework, it isn't
possible to have a pure single-threaded version -- this
single-threaded version uses a Threader class that demands
to use only one thread, and the per-thread variables are
collapsed to a single instance.

Change-Id: Ib49be9ad8ba191bf5f600d8078d1e4e54c97df4d